requirements/base.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.23
django-autoslug==1.9.9
django-cities-light==3.8.1
django-cors-headers==4.8.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
idna==3.10
pillow==11.3.0
progressbar2==4.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
python-utils==3.9.1
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
